Completed search functionality.
There are still a few tweaks that need to be done. For instance, UP
navigation for the chirp details from the search results needs to be
done correctly. Also, the stop words for the pre-processing need to be
loaded in from a file of a bunch of stop words later, instead of the
3-4 that I have in an array.